[v0.30] Allow users to select backend from build script
Before we had AMD support, CMakeLists.txt tried to enable all backends
by default. Now that we have AMD support, that does not make that much
sense so instead it will only enable the backend specified by the user
(with the -DENABLE_XXX_BACKEND flags)

Then, before AMD support, the build.sh script was useful to just
invoke cmake and let it figure out the backends, but the script was
a bit useless after the mentioned change in the CMakeLists.txt.

Therefore, this commit allow users to specify an argument, like:

./build.sh cuda

To specify what backend/s to enable, without the need to manually
configure the build with the -DENABLE_XXX_BACKEND flag. Note that
multiple backends are also allowed, like:

./build.sh intel,hsa

Would enable both Intel and HSA backends (which could make sense for
example in a system with Intel iGPU an an AMD dGPU).